Cf 997 addressbar (#890)
* added withTagsFromURL hoc

* support range toggling in withTagsFromUrl.hoc

* support for partial updates of ranges

* added addTag and removeTag and adjusted range update behaviour a bit

* remove old code

* fix lint errors and stuff

* nock suggester health check

* get rid of react warnings
