+++ src/client/components/filter/FilterCards.component.js
@@ -5,8 +5,9 @@ import Icon from '../base/Icon';
 import T from '../base/T';
 import CardList from './templates/CardList.component.js';
 import CardRange from './templates/CardRange.component.js';
+import withTagsFromUrl from '../base/AdressBar/withTagsFromUrl.hoc';
 
-import './Filters.css';
+import './FilterCards.css';
 
 const params = {
   containerClass: 'Filters__swiper-container',
@@ -92,7 +93,7 @@ class Filters extends React.Component {
     super();
     this.state = {
       showDimmer: false,
-      oFilters: props.cards
+      oFilters: props.filterCards
     };
   }
 
@@ -166,6 +167,7 @@ class Filters extends React.Component {
                 />
               );
             }
+            return null;
           })}
           <div className="FilterCard__space" />
         </Swiper>
@@ -179,4 +181,4 @@ class Filters extends React.Component {
   }
 }
 
-export default Filters;
+export default withTagsFromUrl(Filters);
