+++ src/client/components/filter/templates/CardRange.component.js
@@ -1,13 +1,11 @@
 import React from 'react';
-import {connect} from 'react-redux';
 import {Range} from 'rc-slider';
 import {isEqual} from 'lodash';
 
 import Heading from '../../base/Heading';
 import Icon from '../../base/Icon';
-
-import {filtersMapAll, TOGGLE_FILTER} from '../../../redux/filter.reducer';
-import {getTagsFromUrl} from '../../../redux/selectors';
+import withTagsFromUrl from '../../base/AdressBar/withTagsFromUrl.hoc';
+import {filtersMapAll} from '../../../redux/filter.reducer';
 
 import 'rc-slider/assets/index.css';
 import './CardRange.css';
@@ -26,12 +24,12 @@ class CardRange extends React.Component {
   }
 
   componentWillReceiveProps(nextProps) {
-    if (this.props.selectedTagIds !== nextProps.selectedTagIds) {
-      this.initValue(nextProps.selectedFilters);
+    if (this.props.tags !== nextProps.tags) {
+      this.initValue(nextProps.tags);
     }
   }
 
-  initValue(selectedFilters = this.props.selectedFilters) {
+  initValue(selectedFilters = this.props.tags) {
     const range = this.props.filter.range;
 
     let min = 0;
@@ -39,15 +37,18 @@ class CardRange extends React.Component {
 
     /* if value in selected filter is a range and exist in current range array */
     selectedFilters.forEach(filter => {
-      if (filter instanceof Array) {
-        if (range.includes(filter[0].id) && range.includes(filter[1].id)) {
-          min = range.indexOf(filter[0].id);
-          max = range.indexOf(filter[1].id);
-        }
+      if (
+        filter.type === 'TAG_RANGE' &&
+        range.includes(filter.left.id) &&
+        range.includes(filter.right.id)
+      ) {
+        min = range.indexOf(filter.left.id);
+        max = range.indexOf(filter.right.id);
       }
     });
 
     this.handleChange([min, max]);
+    this.setState({initValue: [min, max]});
   }
 
   getIdByValue(value = this.state.value) {
@@ -69,8 +70,7 @@ class CardRange extends React.Component {
     const oldIds = this.getIdByValue(this.state.initValue);
     const ids = this.getIdByValue(value);
     if (!isEqual(oldIds, ids)) {
-      this.props.toggleFilter(ids);
-      this.setState({initValue: value});
+      this.props.addTag(ids);
     }
   }
 
@@ -129,26 +129,4 @@ class CardRange extends React.Component {
     );
   }
 }
-
-const mapStateToProps = state => {
-  const selectedTagIds = getTagsFromUrl(state);
-  return {
-    selectedTagIds,
-    selectedFilters: selectedTagIds.map(tag => {
-      if (tag instanceof Array) {
-        return tag.map(aTag => filtersMapAll[aTag]);
-      }
-      return filtersMapAll[tag.id || tag];
-    })
-  };
-};
-export const mapDispatchToProps = dispatch => ({
-  toggleFilter: id => dispatch({type: TOGGLE_FILTER, id}),
-  history: (type, path, params = {}) => {
-    dispatch({type, path, params});
-  }
-});
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(CardRange);
+export default withTagsFromUrl(CardRange);
